Remove storybook and use documentation website instead
Storybook is too much of an headache to manage. It often breaks and
requires a custom configuration. Therefore we removed
storybook and decided to use our documentation website for dev
development as well.

Thus we change the build of the documentation website to use our new
meta package @bitflow/bits. The index.ts and package.json of this
package is auto-generated, when a new bit is added. Therefore the
website will always be up to date. Bit authors have only to worry about
changing the useInformation hook of their package for displaying
information on the website.

This will hopefully also improve our documentation.